requirements.txt
Adafruit-Blinka==8.13.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-pm25==2.1.13
adafruit-circuitpython-requests==1.12.12
adafruit-circuitpython-seesaw==1.12.0
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.39.0
Adafruit-PureIO==1.1.9
amqp==5.1.1
appdirs==1.4.4
arandr==0.1.10
asgiref==3.6.0
astroid==2.5.1
asttokens==2.0.4
async-timeout==4.0.2
automationhat==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.4
blinkt==0.1.2
bme680==1.1.1
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorzero==1.1
configparser==5.3.0
cron-descriptor==1.2.35
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
distlib==0.3.1
distro==1.5.0
Django==4.2
django-timezone-field==5.0
djangorestframework==3.14.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.0.12
Flask==1.1.2
fourletterphat==0.1.0
gpiozero==1.6.2
gyp==0.1
html5lib==1.1
i2cdevice==0.0.7
icalendar==4.0.3
idna==2.10
importlib-metadata==1.6.0
inkyphat==1.0.1
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.2.4
lazy-object-proxy==0.0.0
ledshim==0.0.2
lockfile==0.12.2
logilab-common==1.8.1
ltp305==0.0.1
ltr559==0.1.1
lxml==4.6.3
MarkupSafe==1.1.1
matrix11x7==0.0.1
mccabe==0.6.1
microdotphat==0.2.2
more-itertools==4.2.0
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.19.5
oauthlib==3.1.0
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
pimoroni-bme280==0.1.1
pimoroni-ioexpander==0.0.3
pimoroni-mics6814==0.0.2
pimoroni-sgp30==0.0.2
prompt-toolkit==3.0.38
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pyftdi==0.54.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-crontab==2.7.1
python-daemon==3.0.1
python-dateutil==2.8.1
python-prctl==1.7
pytz==2023.3
pyusb==1.2.1
qrcode==6.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scd4x==0.0.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
smbus==1.1.post2
smbus2==0.4.2
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
sqlparse==0.4.4
ssh-import-id==5.10
supervisor==4.2.2
sysv-ipc==1.1.0
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.4.0
tzdata==2023.3
ufw==0.36
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
veml6075==0.0.2
vine==5.0.0
virtualenv==20.4.0+ds
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.0.0
